Change 304 to short test; check $uncompressed defined; dump env
The test for no body with a 304 error message is loosened to only
check that the body is very short (under 10 chars).

The check for the uncompressed body first of all checks that the body
is defined as a new test, then proceeds to check the compression
worked only if it is defined.

An undocumented variable {dump_env} is added to the body to dump the
environment.